refactor: make colors as constant
